add mesh color preferences
this currently uses the same color palettes as the annotation color palette preference. If we want to change this, I can go in and add new images + change the mesh color palette dict in neuroglancer.py. Also, this currently picks colors in order from the palette -- we might want to choose randomly depending on how many mesh layers are typical? or if there tend to be a ton of meshes we might need to choose palettes that have more than 5 colors.